+++ js/topIndianMovies.js
@@ -1,3 +1,4 @@
+import {njs, tmdb, img, v3, v3key,backdrop} from '../js/serverDetails.js'
 var qs= (new URL(document.location)).searchParams;
 var pagenum = Number(qs.get('page')); 
 var count=0;
@@ -19,7 +20,7 @@ $(document).ready(() => {
         $('.pagination #next').attr("href","TopIndianMovies.html?page="+(pagenum+1));
     else
         $('.pagination #next').addClass('disabled');
-    movies = $.get(hostname+"topindian",);
+    var movies = $.get(hostname+"topindian",);
     movies.done((data) => {
         $.each(data, (i,item) => {
             if(count>=start & count <end){
@@ -67,6 +68,42 @@ if(localStorage.getItem('token') == null){
                         $('.auth').hide();
                         $('.authenticated').show();
                         localStorage.setItem('token',msg.token);
+                        location.reload();
+                    }
+                },
+                error: (err)=>{
+                    alert('Invalid Credentials');
+                    console.log(err);
+                }
+            });
+        }
+        else{
+            alert("Enter valid data");
+        }
+    });
+    $('#signup').on('click',()=>{
+        var name = $('[name="name"]').val();
+        var email = $('[name="email"]').val();
+        var uname = $('[name="runame"]').val();
+        var password = $('[name="rpassword"]').val();
+        var password2 = $('[name="password2"]').val();
+        console.log(JSON.stringify({"name":name,"email":email,"username":uname,"password":password}));
+        if(uname != "" && password != ""){
+            $.ajax({
+                type:'POST',
+                dataType: 'json',
+                contentType:'application/json;charset=utf-8',
+                url:njs+'users/signup',
+                data:JSON.stringify({"name":name,"email":email,"username":uname,"password":password}),
+                success: (msg)=>{
+                    console.log(msg);
+                    if(msg.success == true){
+                        $('#registerModal').modal('toggle');
+                        $('#loginModal').modal('toggle');
+                        //$('.auth').hide();
+                        //$('.authenticated').show();
+                        //localStorage.setItem('token',msg.token);
+                        //location.reload();
                     }
                 },
                 error: (err)=>{
@@ -81,13 +118,32 @@ if(localStorage.getItem('token') == null){
     });
     }
     else{
+        $.ajax({
+            type:'GET',
+            url:njs+'users',
+            beforeSend: function (xhr) {
+                xhr.setRequestHeader('Authorization', 'Bearer '+ localStorage.getItem('token'));
+            },
+            dataType:'json',
+            success:(user)=>{
+                $('.logo .name').append(user.name);
+                console.log(user);
+            },
+            error:(err)=>{
+                console.log(err);
+            }
+        });
         $('.authenticated').show();
         $('.auth').hide();
         $('#logout').click(()=>{
-            console.log(localStorage.getItem('token'));
-            localStorage.clear();
-            $('.auth').show();
-            $('.authenticated').hide();
+            var conf = confirm('Do want to logout?');
+            if(conf){
+                console.log(localStorage.getItem('token'));
+                localStorage.clear();
+                $('.auth').show();
+                $('.authenticated').hide();
+                location.reload();
+            }
         });
     }
