+++ js/topTvShows.js
@@ -1,48 +1,56 @@
+import { key, tmdb, njs, img, cast_img}  from '../js/serverDetails.js';
 var qs= (new URL(document.location)).searchParams;
 var pagenum = Number(qs.get('page')); 
-var imgurl = 'https://image.tmdb.org/t/p/w500';
-var count=0;
 var start =-1;
 var end = 20; 
 if(pagenum > 1){
     end = pagenum * 20;
     start = end-20;
 }
-const hostname='http://localhost:3000/';
 $(document).ready(() => {
     $('.pagination #'+pagenum).addClass('active');
     if(pagenum != 1)
-        $('.pagination #previous').attr("href","topTvShows.html?page="+(pagenum-1));
+        $('.pagination #previous').attr("href","topTvShow.html?page="+(pagenum-1));
     else    
         $('.pagination #previous').addClass('disabled');
     if(pagenum != 10)
-        $('.pagination #next').attr("href","topTvShows.html?page="+(pagenum+1));
+        $('.pagination #next').attr("href","topTvShow.html?page="+(pagenum+1));
     else
         $('.pagination #next').addClass('disabled');
-    movies = $.get(hostname+"toptv",);
-    movies.done((data) => {
-        $.each(data, (i,item) => {
-            if(i>start & i <end){
-                $(".movies-list").append
-                ("<div class='col-12 col-sm-6 movie-tiles pr-5 pl-5 pr-md-2 pl-md-2' height='400px'>"+
-                    "<div class='card movie-card'>"+
-                        "<div class='row'>"+
-                            "<div class='col-12 col-md-5' style='text-align:center'>"+
-                                "<a href='tvPage.html?id="+item._id+"'>"+
-                                "<img class='card-img-top' src='" + imgurl+item.poster_path + "' height='340px' width='240px'><a>"+
-                            "</div>"+
-                            "<div class='col-12 col-md-6 mov-des'>"+
-                                "<h1>"+ item.name +"("+ item.first_air_date.split("-")[0] +")</h1>"+
-                                "<p class='mov-desc' style='padding: 10px'>"+
-                                    item.overview.split('. ')[0] + "<a href='tvPage.html?id="+item._id+"'>......more</a>"+
-                                "</p>"+
-                            "</div>"+
-                        "</div>"+
-                    "</div>"+
-                "</div>");
-            }
-        });
-    });
+
+
+        var tv=$.ajax({
+            url: njs+'toptv',
+            method: 'GET',
+            dataType: 'JSON', 
+          })
+          .done((data) => {
+            console.log(data);
+            $.each(data ,(i,item) =>{
+              if(i>start & i<end){
+                $('.movies').append("<div class='row element'>"+
+                                    "<div class='col-5 col-sm-4 mov-img-container'>"+
+                                    "<a href='tvPage.html?_id="+item._id+"'><img src='"+cast_img+item.poster_path+"' class='img-responsive mov-img'></a>"+
+                                    "</div>"+
+                                    "<div class='col-7 col-sm-8' style='padding: 0px'>"+
+                                      "<div class='container-fluid'>"+
+                                        "<div class='row'>"+
+                                          "<div class='col-12'>"+
+                                            "<h2 class='mov-title'><a href='tvPage.html?_id="+item._id+"'>"+item.name+" </a></h2><h4 class='mov-year'>"+item.first_air_date.split('-')[0]+"</h4>"+
+                                          "</div>"+
+                                          "<div class='col-12'>"+
+                                              "<div class='mov-desc'><p>"+item.overview+"</p></div>"+
+                                          "</div>"+
+                                        "</div>"+
+                                      "</div>"+
+                                    "</div>"+
+                                    "</div><hr class='div-line'>");
+                  }
+              });
+          })
+          .fail(()=>{
+            console.log('Error');
+          });   
 });
 
 
@@ -67,6 +75,42 @@ if(localStorage.getItem('token') == null){
                         $('.auth').hide();
                         $('.authenticated').show();
                         localStorage.setItem('token',msg.token);
+                        location.reload();
+                    }
+                },
+                error: (err)=>{
+                    alert('Invalid Credentials');
+                    console.log(err);
+                }
+            });
+        }
+        else{
+            alert("Enter valid data");
+        }
+    });
+    $('#signup').on('click',()=>{
+        var name = $('[name="name"]').val();
+        var email = $('[name="email"]').val();
+        var uname = $('[name="runame"]').val();
+        var password = $('[name="rpassword"]').val();
+        var password2 = $('[name="password2"]').val();
+        console.log(JSON.stringify({"name":name,"email":email,"username":uname,"password":password}));
+        if(uname != "" && password != ""){
+            $.ajax({
+                type:'POST',
+                dataType: 'json',
+                contentType:'application/json;charset=utf-8',
+                url:njs+'users/signup',
+                data:JSON.stringify({"name":name,"email":email,"username":uname,"password":password}),
+                success: (msg)=>{
+                    console.log(msg);
+                    if(msg.success == true){
+                        $('#registerModal').modal('toggle');
+                        $('#loginModal').modal('toggle');
+                        //$('.auth').hide();
+                        //$('.authenticated').show();
+                        //localStorage.setItem('token',msg.token);
+                        //location.reload();
                     }
                 },
                 error: (err)=>{
@@ -81,13 +125,32 @@ if(localStorage.getItem('token') == null){
     });
     }
     else{
+        $.ajax({
+            type:'GET',
+            url:njs+'users',
+            beforeSend: function (xhr) {
+                xhr.setRequestHeader('Authorization', 'Bearer '+ localStorage.getItem('token'));
+            },
+            dataType:'json',
+            success:(user)=>{
+                $('.logo .name').append(user.name);
+                console.log(user);
+            },
+            error:(err)=>{
+                console.log(err);
+            }
+        });
         $('.authenticated').show();
         $('.auth').hide();
         $('#logout').click(()=>{
-            console.log(localStorage.getItem('token'));
-            localStorage.clear();
-            $('.auth').show();
-            $('.authenticated').hide();
+            var conf = confirm('Do want to logout?');
+            if(conf){
+                console.log(localStorage.getItem('token'));
+                localStorage.clear();
+                $('.auth').show();
+                $('.authenticated').hide();
+                location.reload();
+            }
         });
     }
